Migrando model de bloco de sessao para parlamentar
